ai_engine/requirements.txt
flask==3.0.0
numpy==1.24.3
scikit-image==0.21.0
pillow==10.0.0
requests==2.31.0
flask-cors==4.0.0
gunicorn==21.2.0
